backend/requirements.txt
# === Web Framework ===
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# === Banco de Dados ===
sqlalchemy>=2.0.48
alembic>=1.13.0
aiosqlite>=0.20.0          # SQLite async (dev)
asyncpg>=0.29.0            # PostgreSQL async (prod)
psycopg2-binary>=2.9.9     # PostgreSQL sync (Alembic)

# === Tarefas Assíncronas ===
celery[redis]>=5.3.6
redis>=5.0.3

# === NLP / Similaridade ===
rapidfuzz>=3.0.0
fuzzywuzzy>=0.18.0

# === Autenticação ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0          # Fernet para tokens OAuth + transitiva de jose

# === Configuração ===
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.0

# === HTTP ===
httpx>=0.27.0
requests>=2.31.0

# === IA / LLM ===
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
anthropic>=0.25.0
openai>=1.16.0

# === Redes Sociais ===
tweepy>=4.14.0
apify-client>=1.6.0

# === Email ===
resend>=0.7.0

# === Google APIs ===
google-api-python-client>=2.127.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# === Geração de Imagens / Carrossel ===
Pillow>=10.3.0
jinja2>=3.1.3
playwright>=1.43.0

# === Testes ===
pytest>=8.1.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.27.0              # TestClient assíncrono

# === Observabilidade ===
sentry-sdk[fastapi]>=2.0.0
structlog>=24.1.0

# === Qualidade de código ===
black>=24.4.0
ruff>=0.4.0
bandit>=1.7.8

# === Profiling ===
pyinstrument>=4.6.0
memory-profiler>=0.61.0
py-spy>=0.3.14
